# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  chat_sdk.CreateGroupRoomReq:
    properties:
      members:
        items:
          type: integer
        type: array
      name:
        type: string
    required:
    - members
    - name
    type: object
...
  chat_sdk.RoomMemberReq:
    properties:
      room_id:
        example: 1
        type: integer
      user_id:
        description: remove 用
        example: 1001
        type: integer
      user_ids:
        description: add 用（批量）
        example:
        - 1001
        items:
          type: integer
        type: array
    required:
    - room_id
    type: object
...
  chat_sdk.SetGroupAdminReq:
    properties:
      is_admin:
        type: boolean
      room_id:
        type: integer
      target_user_id:
        type: integer
    required:
    - room_id
    - target_user_id
    type: object
  chat_sdk.SetGroupMuteReq:
    properties:
      duration_minutes:
        description: 0 to cancel
        type: integer
      room_id:
        type: integer
    required:
    - room_id
    type: object
  chat_sdk.SetGroupMuteScheduledReq:
    properties:
      duration_minutes:
        type: integer
      room_id:
        type: integer
      start_time:
        description: HH:MM
        type: string
    required:
    - duration_minutes
    - room_id
    - start_time
    type: object
  chat_sdk.SetMyGroupNicknameReq:
    properties:
      nickname:
        example: 我在群里的昵称
        type: string
      room_id:
        example: 1
        type: integer
    required:
    - room_id
    type: object
  chat_sdk.SetUserMuteReq:
    properties:
      duration_minutes:
        description: 0 to cancel
        type: integer
      room_id:
        type: integer
      target_user_id:
        type: integer
    required:
    - room_id
    - target_user_id
    type: object
  chat_sdk.UpdateGroupInfoReq:
    properties:
      avatar:
        type: string
      name:
        type: string
      room_id:
        type: integer
    required:
    - room_id
    type: object
...
  models.Room:
    properties:
      avatar:
        description: 房间头像
        type: string
      createdAt:
        type: string
      creator:
        allOf:
        - $ref: '#/definitions/models.User'
        description: 关联关系
      creatorID:
        description: 创建者 ID
        type: integer
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        description: 描述
        type: string
      id:
        type: integer
      isEncrypted:
        description: 是否端到端加密
        type: boolean
      isMute:
        description: 新增禁言相关字段
        type: boolean
      memberLimit:
        description: 成员上限
        type: integer
      messages:
        items:
          $ref: '#/definitions/models.Message'
        type: array
      muteDailyDuration:
        description: 每日禁言持续时长（分钟）
        type: integer
      muteDailyStartTime:
        description: 每日禁言开始时间 "HH:MM"
        type: string
      muteUntil:
        description: 全员禁言截止时间（倒计时模式）
        type: string
      name:
        description: 房间名称
        type: string
      roomAccount:
        description: |-
          RoomAccount 对外房间号/群号（可用 10 位数字字符串或自定义规则），用于搜索/分享；
          不参与任何外键关联，避免再被 GORM 推断成 bigint。
        type: string
      roomUsers:
        items:
          $ref: '#/definitions/models.RoomUser'
        type: array
      type:
        description: '类型: 1-私聊 2-群聊'
        type: integer
      updatedAt:
        type: string
    type: object
...
  models.User:
    properties:
      avatar:
        description: 头像
        type: string
      birthday:
        description: 生日
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        description: 邮箱
        type: string
      friendApplys:
        items:
          $ref: '#/definitions/models.FriendApply'
        type: array
      friends:
        description: 关联关系
        items:
          $ref: '#/definitions/models.Friend'
        type: array
      gender:
        description: '性别: 0-未知 1-男 2-女'
        type: integer
      id:
        type: integer
      lastActiveAt:
        description: 最后活跃时间
        type: string
      lastLoginAt:
        description: 最后登录时间
        type: string
      messages:
        items:
          $ref: '#/definitions/models.Message'
        type: array
      nickname:
        description: 昵称
        type: string
      onlineStatus:
        description: '在线状态: 0-离线 1-在线'
        type: integer
      password:
        description: 密码
        type: string
      phone:
        description: 手机号
        type: string
      rooms:
        items:
          $ref: '#/definitions/models.RoomUser'
        type: array
      signature:
        description: 个性签名
        type: string
      uid:
        description: 对外用户 ID
        type: string
      updatedAt:
        type: string
      username:
        description: 用户名
        type: string
    type: object
  response.Response:
    properties:
      code:
        description: 业务状态码
        example: 0
        type: integer
      data:
        description: 响应数据
        type: object
      msg:
        description: 提示消息
        example: success
        type: string
    type: object
...
  service.GroupInfoDTO:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      creator_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      room_account:
        type: string
      updated_at:
        type: string
    type: object
...
  service.RoomDTO:
    properties:
      avatar:
        type: string
      id:
        type: integer
      last_message:
        $ref: '#/definitions/service.MessageDTO'
      name:
        type: string
      room_account:
        type: string
      type:
        type: integer
      unread_count:
        type: integer
      updated_at:
        type: string
    type: object
  service.RoomMemberListItemDTO:
    properties:
      avatar:
        type: string
      display_name:
        type: string
      group_nickname:
        description: 群昵称（room_user.nickname）
        type: string
      is_muted:
        type: boolean
      nickname:
        type: string
      remark:
        description: 好友备注（当前用户视角）
        type: string
      role:
        type: integer
      user_id:
        type: integer
      username:
        type: string
    type: object
...
host: localhost:6789
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/cydxin/chat-sdk/issues
  description: |-
    即时通讯 SDK 的 RESTful API 文档，包含用户管理、消息、房间、好友等模块

    ## 业务状态码说明
    | Code | 说明 |
    |------|------|
    | 0 | 成功 |
    | 10001 | 参数错误 |
    | 10002 | 用户不存在 |
    | 10003 | 密码错误（登录失败） |
    | 10004 | Token 无效 |
    | 10005 | 权限不足 |
    | 99999 | 内部错误 |

    ## HTTP 状态码说明
    - **200**: 业务请求成功（根据 response.code 判断业务状态）
    - **401**: 认证失败（未登录/Token 无效/登录失败）
    - **403**: 权限不足
    - **500**: 服务器内部错误

    ## 响应格式
    所有接口统一返回格式：
    ```json
    {
    "code": 0,
    "msg": "success",
    "data": {}
    }
    ```
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/cydxin/chat-sdk
  title: Chat SDK API
  version: "1.0"
paths:
...
  /room/admin/set:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.SetGroupAdminReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 设置管理员
      tags:
      - Room
  /room/group:
    post:
      consumes:
      - application/json
      description: 创建新的群聊房间
      parameters:
      - description: 创建参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.CreateGroupRoomReq'
      produces:
      - application/json
      responses:
        "200":
          description: 房间信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Room'
              type: object
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建群聊
      tags:
      - 房间
  /room/group/info:
    get:
      consumes:
      - application/json
      description: 根据 room_id 获取群聊基础信息（不含成员列表）
      parameters:
      - description: 群ID(房间ID)
        format: int64
        in: query
        name: room_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 群信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.GroupInfoDTO'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取群基础信息
      tags:
      - 房间
  /room/group/list:
    get:
      consumes:
      - application/json
      description: 获取当前用户参与的所有群聊（仅 Type=2）
      produces:
      - application/json
      responses:
        "200":
          description: 群聊列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.RoomDTO'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户群聊列表
      tags:
      - 房间
  /room/group/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.UpdateGroupInfoReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新群信息
      tags:
      - Room
  /room/list:
    get:
      consumes:
      - application/json
      description: 获取当前用户参与的所有房间
      produces:
      - application/json
      responses:
        "200":
          description: 房间列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.RoomDTO'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户房间列表
      tags:
      - 房间
  /room/member/add:
    post:
      consumes:
      - application/json
      description: 将用户添加到房间
      parameters:
      - description: 成员信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.RoomMemberReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 添加房间成员
      tags:
      - 房间
  /room/member/check:
    get:
      consumes:
      - application/json
      description: 检查用户是否是房间成员，如果不传 user_id 则检查当前用户
      parameters:
      - description: 房间ID
        format: int64
        in: query
        name: room_id
        required: true
        type: integer
      - description: 用户ID (不传则查自己)
        format: int64
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 检查结果
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: boolean
                  type: object
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 检查房间成员
      tags:
      - 房间
  /room/member/list:
    get:
      consumes:
      - application/json
      description: 获取指定房间(群)成员列表，展示名按：好友备注 > 群昵称 > 用户昵称 > 用户名
      parameters:
      - description: 房间ID
        format: int64
        in: query
        name: room_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成员列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.RoomMemberListItemDTO'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取群成员列表
      tags:
      - 房间
  /room/member/nickname:
    post:
      consumes:
      - application/json
      description: 设置当前用户在指定房间(群)中的昵称（room_user.nickname），仅影响自己视角
      parameters:
      - description: 请求参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.SetMyGroupNicknameReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 设置群昵称
      tags:
      - 房间
  /room/member/remove:
    post:
      consumes:
      - application/json
      description: 将用户从房间移除
      parameters:
      - description: 成员信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.RoomMemberReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 移除房间成员
      tags:
      - 房间
  /room/mute/group:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.SetGroupMuteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 设置群禁言（倒计时）
      tags:
      - Room
  /room/mute/group/scheduled:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.SetGroupMuteScheduledReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 设置群禁言（定时）
      tags:
      - Room
  /room/mute/user:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.SetUserMuteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 设置用户禁言
      tags:
      - Room
  /room/private:
    post:
      consumes:
      - application/json
      description: 创建或获取两人私聊房间
      parameters:
      - description: 目标用户ID
        format: int64
        in: query
        name: target_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 房间信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Room'
              type: object
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建私聊
      tags:
      - 房间
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 格式：Bearer <token>
    in: header
    name: Authorization
    type: apiKey
  QueryToken:
    description: 用于 WebSocket 等无法传 header 的场景
    in: query
    name: token
    type: apiKey
swagger: "2.0"